+++ src/halomod/bias.py
@@ -67,8 +67,10 @@ from hmf.cosmology.cosmo import astropy_to_colossus
 from colossus.lss.bias import haloBiasFromNu
 from astropy.cosmology import FLRW, Planck15
 from hmf.halos.mass_definitions import SOMean
+from hmf._internals import pluggable
 
 
+@pluggable
 class Bias(Component):
     r"""
     The base Bias component.
@@ -157,10 +159,6 @@ class Bias(Component):
         >>> plt.plot(peak_height, bias.bias())
         """
         return np.ones_like(self.nu)
-
-    def __init_subclass__(cls, **kwargs):
-        super().__init_subclass__(**kwargs)
-        cls._models[cls.__name__] = cls
 
 
 class UnityBias(Bias):
@@ -742,6 +740,7 @@ class Tinker10PBSplit(Bias):
         )
 
 
+@pluggable
 class ScaleDepBias(Component):
     r"""Base class for scale-dependent bias models.