+++ CHANGELOG.rst
@@ -1,8 +1,11 @@
 Releases
 ========
 
-2.0.0rc1
---------
+2.0.0 [25th Nov 2020]
+---------------------
+This is a **major** new release that brings halomod into the properly
+Python 3 world, and tightens up a lot of issues. It also corresponds to the
+submission of `Murray, Diemer and Chen (2020) <https://arxiv.org/abs/2009.14066>`_.
 
 Features
 ++++++++
@@ -11,20 +14,22 @@ Features
 * ExtendedSpline class which does better (and more customized) extrapolation on splines.
 * Added ability to change the underlying distribution of satellite and central counts.
 * Python 3!
-* New UnityBias component
+* New ``UnityBias`` component
 * Ability to do cross-correlations between one tracer and another
 * Delineation between DM and halos in power spectra
 * Much more documentation
-* Addition of CoredNFW profile
+* Addition of ``CoredNFW`` profile
 * Added ``mean_tracer_den_unit`` -- the mean density of whatever tracer you use.
 * Interfaces with COLOSSUS for concentration and bias.
 * New CLI interface ``halomod run``
 * New examples, including an example of parameter fitting with ``halomod`` and ``emcee`.
+* New hod, density profile and concentration model, with new example, for HI.
 
 Changes
 +++++++
 
 * Removed dynamic ``mmin``.
+* No more support for Python 2.
 
 Bugfixes
 ++++++++
@@ -33,7 +38,12 @@ Bugfixes
 * Fix for when no tracer profile/concentration params are given. Now use the halo params
   iff the tracer model is equal to the halo one.
 * Fix for halo profile numerical FT.
+* Fix #59 -- can't have non-dict ``model_params``.
 * Fix for calculation of halo centre power spectrum
+* Fix for tracer params initialisation.
+* Fix amplitude of ``angular_corr_gal`` when using ``p_of_z=True``
+  (`Issue #63 <https://github.com/steven-murray/halomod/issues/63>`_,
+   `PR #72 <https://github.com/steven-murray/halomod/pull/72>`_)
 
 v1.4.5
 ------